Print ts model plts

print_ts_mdl_plts() is a Print function that prints output to console Specifically, this function implements an algorithm to print ts model plts. The function is called for its side effects and does not return a value.

print_ts_mdl_plts(paths_to_plts_chr, title_1L_chr, label_refs_chr, mdl_smry_ls)

Arguments

paths_to_plts_chr
Paths to plts (a character vector)

title_1L_chr
Title (a character vector of length one)

label_refs_chr
Label references (a character vector)

mdl_smry_ls
Model smry (a list)